test: reorganize and expand coverage
